apps/backend/services/auth/requirements.txt
# Core Framework
fastapi
uvicorn[standard]

# Data Validation & Settings
pydantic
pydantic-settings

# Database (ORM & Driver)
sqlalchemy[asyncio]
asyncpg
alembic

# Security & Authentication
passlib[bcrypt]
python-jose[cryptography]
pyotp
qrcode[pil]

# Utilities
python-dotenv
httpx
python-json-logger
PyYAML

# Redis for caching and rate limiting
redis[hiredis]
aioredis
